align cpu count
