+++ packages/renderer/src/get-concurrency.ts
@@ -1,32 +1,9 @@
-import {execSync} from 'node:child_process';
-import os from 'node:os';
-
-// Kubernetes uses the following command to spawn Docker containers:
-// docker run --cpuset-cpus="0,1" to assign only 2 CPUs.
-// However, Node.js returns the core count of the host system (up to 96!)
-// We also get it from nproc and use the minimum of the two.
-const getConcurrencyFromNProc = (): number | null => {
-	try {
-		return parseInt(execSync('nproc').toString().trim(), 10);
-	} catch (error) {
-		return null;
-	}
-};
-
-const getMaxCpus = () => {
-	const node = os.cpus().length;
-	const nproc = getConcurrencyFromNProc();
-	if (nproc === null) {
-		return node;
-	}
-
-	return Math.min(nproc, node);
-};
+import {getCpuCount} from './get-cpu-count';
 
 export const getActualConcurrency = (
 	userPreference: number | string | null,
 ) => {
-	const maxCpus = getMaxCpus();
+	const maxCpus = getCpuCount();
 
 	if (userPreference === null) {
 		return Math.round(Math.min(8, Math.max(1, maxCpus / 2)));
