+++ packages/renderer/src/get-video-threads-flag.ts
@@ -1,10 +1,11 @@
 import os from 'node:os';
+import {getCpuCount} from './get-cpu-count';
 const MEMORY_USAGE_PER_THREAD = 400_000_000; // 400MB
 const RESERVED_MEMORY = 2_000_000_000;
 
 export const getIdealVideoThreadsFlag = () => {
 	const freeMemory = os.freemem();
-	const cpus = os.cpus().length;
+	const cpus = getCpuCount();
 
 	const maxRecommendedBasedOnCpus = (cpus * 2) / 3;
 	const maxRecommendedBasedOnMemory =
